pyproject.toml
# Packaging instructions for ticc-clustering

### Copyright 2024 National Technology & Engineering Solutions of Sandia,
### LLC (NTESS). Under the terms of Contract DE-NA0003525 with NTESS, the
### U.S. Government retains certain rights in this software.
###
### Redistribution and use in source and binary forms, with or without
### modification, are permitted provided that the following conditions are
### met:
###
### 1. Redistributions of source code must retain the above copyright
###    notice, this list of conditions and the following disclaimer.
###
### 2. Redistributions in binary form must reproduce the above copyright
###    notice, this list of conditions and the following disclaimer in
###    the documentation and/or other materials provided with the
###    distribution.
###
### 3. Neither the name of the copyright holder nor the names of its
###    contributors may be used to endorse or promote products derived
###    from this software without specific prior written permission.
###
### THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
### “AS IS” AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
### LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
### A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
### HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
### SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
### LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
### DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
### THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
### (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
### OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.


[tool.poetry]

name = "metric-dbscan"
version = "1.0.1"
description = "DBSCAN for general metric spaces"
license = "BSD-3-Clause"

packages = [
    { include = "metric_dbscan", from = "src" }
]

authors = [
    "Andy Wilson <atwilso@sandia.gov>",
    "Keith Dalbey <kdalbey@sandia.gov>",
    "Daniel DeLayo <drdelay@sandia.gov>",
    "Renee Gooding <rlgoodn@sanida.gov>",
    "Jessica Jones <jjones7@sandia.gov>",
    "Kanad Khanna <kkhanna@sandia.gov>",
    "Nitin Sharan <nsharan@sandia.gov>",
    "Jon Whetzel <jhwhetz@sandia.gov"
]

maintainers = [
    "Andy Wilson <atwilso@sandia.gov>"
]

readme = "README.md"
repository = "https://github.com/sandialabs/metric-dbscan"
keywords = [
    "clustering",
    "dbscan",
    "metric space"
]

classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: BSD License",
    "Programming Language :: Python",
    "Topic :: Scientific/Engineering :: Information Analysis"
]


[tool.poetry.dependencies]
python = ">=3.6"
tqdm = ">=2.0"

[tool.poetry.group.test.dependencies]
python-levenshtein = "*"

[build-system]
requires = [ "poetry-core" ]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib"
]
pythonpath = "src"
